✅ splits up common gql test code to util to avoid redundant tests
