+++ gqlValidation/queryValidations.test.ts
@@ -5,29 +5,27 @@ import { validate, specifiedRules, NoUnusedFragmentsRule } from 'graphql/validat
 import path from 'path';
 import { expect } from 'chai';
 import urlJoin from 'url-join';
-import { compileGqlAst, getGqlFiles, ABSOLUTE_ROOT_DIR } from './compilation.test';
+import { ABSOLUTE_ROOT_DIR, compileGqlAst, getGqlFiles } from './utils';
 
 require('dotenv').config();
 
 const GRAPHQL_URL = urlJoin(process.env.GATEWAY_API_ROOT, 'graphql');
 const validationRules = specifiedRules.filter(rule => rule !== NoUnusedFragmentsRule);
 
-describe('compatible', async () => {
-  it(`must be compatible`, async () => {
+describe('this test', async () => {
+  let gqlFiles;
+  it('can get gql files', async () => {
+    gqlFiles = await getGqlFiles();
+  });
+  it(`can retrieve remote schema at ${GRAPHQL_URL}`, async () => {
     const apolloLink = new HttpLink({
       uri: GRAPHQL_URL,
       fetch,
     });
-    let schema;
-    try {
-      schema = await introspectSchema(apolloLink);
-    } catch (err) {
-      throw new Error(`failed to retrieve remote schema`);
-    }
-    const gqlFiles = await getGqlFiles();
+    const schema = await introspectSchema(apolloLink);
     gqlFiles.forEach(({ dir, file }) => {
-      describe(path.resolve(dir, file).replace(ABSOLUTE_ROOT_DIR, ''), () => {
-        it(`must be compliant with gql API at ${GRAPHQL_URL}`, () => {
+      describe(path.resolve(dir, file).replace(ABSOLUTE_ROOT_DIR, 'components'), () => {
+        it(`is compliant with gql API at ${GRAPHQL_URL}`, () => {
           validateGqlFileAgainstSchema(schema)({ dir, file });
         });
       });
@@ -40,7 +38,7 @@ const validateGqlFileAgainstSchema = schema => ({ dir, file }) => {
   const queryAst = compileGqlAst(queryFilePath);
   const errors = validate(schema, queryAst, validationRules);
   if (errors.length) {
-    throw new Error(`validation failed for ${queryFilePath}: ${errors}`);
+    throw new Error(errors);
   }
   expect(errors).to.be.empty;
 };
